fix замечания

+++ controllers/cards.js
@@ -1,6 +1,6 @@
 const { Card } = require('../models');
 const { LIKE_OPTIONS, ERROR_MESSAGE } = require('../constants');
-const { IncorrectDataError, NotFoundError } = require('../helpers/Errors');
+const { IncorrectDataError, NotFoundError, ForbiddenError } = require('../helpers/Errors');
 
 const getAllCards = (req, res, next) => {
   Card.find({})
@@ -13,7 +13,7 @@ const createCard = (req, res, next) => {
   const { name, link } = req.body;
 
   Card.create({ name, link, owner: req.user._id })
-    .orFail(new IncorrectDataError(ERROR_MESSAGE.INCORRECT_DATA))
+    .catch((err) => new IncorrectDataError(`${ERROR_MESSAGE.INCORRECT_DATA}: ${err.message}`))
     .then((card) => res.send({ data: card }))
     .catch(next);
 };
@@ -21,9 +21,16 @@ const createCard = (req, res, next) => {
 const deleteCard = (req, res, next) => {
   const { cardId } = req.params;
 
-  Card.findByIdAndRemove(cardId)
-    .orFail(new IncorrectDataError(ERROR_MESSAGE.INCORRECT_DATA))
-    .then((card) => res.send({ data: card }))
+  Card.findById(cardId).orFail().catch(() => {
+    throw new NotFoundError(ERROR_MESSAGE.CARD_NOT_FOUND);
+  }).then((card) => {
+    if (card.owner.toString() !== req.user._id) {
+      throw new ForbiddenError(ERROR_MESSAGE.NOT_AUTHORIZED);
+    }
+    Card.findByIdAndRemove(cardId)
+      .then((cardData) => res.send({ data: cardData }))
+      .catch(next);
+  })
     .catch(next);
 };
 
@@ -38,7 +45,8 @@ const likeCard = (req, res, next) => {
     },
     LIKE_OPTIONS,
   )
-    .orFail(new NotFoundError(ERROR_MESSAGE.CARD_NOT_FOUND))
+    .orFail()
+    .catch(() => new NotFoundError(ERROR_MESSAGE.CARD_NOT_FOUND))
     .then((likes) => res.send({ data: likes }))
     .catch(next);
 };
@@ -54,7 +62,8 @@ const dislikeCard = (req, res, next) => {
     },
     LIKE_OPTIONS,
   )
-    .orFail(new NotFoundError(ERROR_MESSAGE.CARD_NOT_FOUND))
+    .orFail()
+    .catch(() => new NotFoundError(ERROR_MESSAGE.CARD_NOT_FOUND))
     .then((likes) => res.send({ data: likes }))
     .catch(next);
 };
